#  storage/ api-docs/ api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api Project",
    "description": "Demo my Project ",
    "contact": {
      "email": "admin@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login",
        "requestBody": {
          "description": "User login data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "d764dd091cc4494ae0baf360b03319f3",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "email",
                  "lastName",
                  "name",
                  "phone",
                  "image",
                  "password",
                  "password_confirmation"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "email": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "password_confirmation": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add Category."
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "69281b12abb272c76871f19cb17ca563",
        "responses": {
          "200": {
            "description": "Display a listing of projects."
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
        "responses": {
          "200": {
            "description": "Display a listing of projects."
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
    "/api/auth/user-profile": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "4188c99fa2637f92b254e6526b44fc50",
        "responses": {
          "200": {
            "description": "Display a listing of projects."
          }
        },
        "security": [
          {
            "apiAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  }
}